backend/requirements_minimal.txt
# Core FastAPI packages
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.23
databases[sqlite]>=0.8.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP & File handling
httpx>=0.25.2
python-multipart>=0.0.6

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Optional: OCR (only install if needed)
# easyocr>=1.7.2
# pillow>=10.0.0

# WebSocket
websockets>=12.0

# Environment
python-dotenv>=1.0.0